Scala 3 support (#3633)
* Scala 3 project xcompile setup and first modules compat

* monocle 3.3 upd

* vector cross compile

* raster cross compile

* scala3 ci builds

* chronoscala update

* shapefile cross compile

* gdal, geotools, vectortile cross compile

* layer cross compile

* store cross compile

* reorder scala3 modules runs

* s3 cross compile

* accumulo cross compile

* cassandra cross compile

* hbase cross compile

* spark cross compile

* spark-testkit cross compile

* spark cross compile

* s3-spark cross compile

* accumulo-spark cross compile

* cassandra-spark cross compile

* gdal-spark cross compile

* bench cross compile

* hbase-spark cross compile

* jmh fix

* pipeline module circe cleanup

* pipeline module cross compile and doc-examples

* rely on the full cross build now

* MiMa readjustments

* build cleanup

+++ s3-spark/src/main/scala/geotrellis/spark/store/s3/S3RDDWriter.scala
@@ -110,7 +110,7 @@ class S3RDDWriter(
             val request = PutObjectRequest.builder()
               .bucket(bucket)
               .key(key)
-              .contentLength(contentBytes.length)
+              .contentLength(contentBytes.length.toLong)
               .build()
             val requestBody = RequestBody.fromBytes(contentBytes)
